* dev/11/leo:
  整理 API
  Check Delete DNS Pod Count is Match
  Modify Delete Line
  add Seed by ScanPlatform Seeder
  add Seed by ScanPlatform Seeder

+++ app/Http/Controllers/Api/v1/LineController.php
@@ -3,17 +3,20 @@
 namespace App\Http\Controllers\Api\v1;
 
 use App\Http\Controllers\Controller;
-use Hiero7\Enums\InputError;
-use Hiero7\Models\LocationNetwork as Line;
-use Hiero7\Services\LineService;
 use App\Http\Requests\LineRequest as Request;
+use Hiero7\Enums\InputError;
+use Hiero7\Enums\InternalError;
+use Hiero7\Models\LocationNetwork as Line;
+use Hiero7\Services\DnsProviderService;
+use Hiero7\Services\LineService;
 use Hiero7\Services\SchemeService;
+use Illuminate\Support\Collection;
 
 class LineController extends Controller
 {
-
     protected $lineService;
     protected $schemeService;
+
     /**
      * LineController constructor.
      */
@@ -31,8 +34,10 @@ class LineController extends Controller
 
     public function create(Request $request)
     {
-        $request->merge(['edited_by' => $this->getJWTPayload()['uuid']]);
+        $this->requestMergeEditedBy($request);
+
         $errorCode = null;
+
         $line = [];
         if ($this->lineService->checkNetworkId($request->get('network_id'))) {
             $errorCode = InputError::THE_NETWORK_ID_EXIST;
@@ -49,14 +54,44 @@ class LineController extends Controller
 
     public function edit(Request $request, Line $line)
     {
-        $request->merge(['edited_by' => $this->getJWTPayload()['uuid']]);
+        $this->requestMergeEditedBy($request);
+
         $line->update($request->only('continent_id', 'country_id', 'location', 'isp', 'mapping_value'));
+
         return $this->response("Success", null, $line);
     }
 
-    public function destroy(Line $line)
+    public function destroy(Line $line, DnsProviderService $dnsProviderService)
     {
+        $deleteData = $line->locationDnsSetting->map(function ($locationDnsSetting) {
+            return ['id' => $locationDnsSetting->provider_record_id];
+        });
+
+        if (!$this->deletDNSRecord($dnsProviderService, $deleteData)) {
+            return $this->setStatusCode(409)->response('please contact the admin', InternalError::INTERNAL_ERROR);
+        }
+
+        $line->locationDnsSetting()->delete();
+
         $line->delete();
+
         return $this->response();
+    }
+
+    private function deletDNSRecord(DnsProviderService $dnsProviderService, Collection $deleteData): bool
+    {
+        if (!$deleteData->count()) {
+            return true;
+        }
+
+        $response = $dnsProviderService->syncRecordToDnsPod([
+            'delele' => json_encode($deleteData),
+        ]);
+
+        if ($response && !$response['errorCode']) {
+            return ($deleteData->count() == count($response['data']['deleteSync'])) ? true : false;
+        }
+
+        return false;
     }
 }
